- Realised the documentation uses long for ids, changed the code to use long. - Added a response body for when adding a new vote - created a method to find results for dates that suit all voters (For now method is rather heavy processing wise, should be refactored if there is time) - renamed a couple dto to exclude request and response in name.

+++ sql/schema.sql
@@ -1,23 +1,23 @@
 create table event (
-    id int auto_increment primary key,
+    id bigint auto_increment primary key,
     name varchar(255) not null
 );
 
 create table event_date (
-    id int auto_increment primary key,
-    event_id int not null,
+    id bigint auto_increment primary key,
+    event_id bigint not null,
     date date not null,
     foreign key (event_id) references event(id)
 );
 
 create table person (
-    id int auto_increment primary key,
+    id bigint auto_increment primary key,
     name varchar(255) not null
 );
 
 create table event_date_vote (
-    event_date_id int not null,
-    voter_id int not null,
+    event_date_id bigint not null,
+    voter_id bigint not null,
     primary key (event_date_id, voter_id),
     foreign key (event_date_id) references event_date(id),
     foreign key (voter_id) references person(id)
